+++ src/app/[lang]/layout.tsx
@@ -1,5 +1,6 @@
 import Navbar from "@/components/Navbar";
 import Content from "@/components/Content";
+import FooterWrapper from "@/components/FooterWrapper";
 import Script from "next/script";
 
 import { getUser } from "@/lib/data";
@@ -43,10 +44,10 @@ export default async function Root({ params, children }: Props) {
             `
           }}
         />
-      </head>
-      <body className="relative min-h-screen overflow-y-auto bg-slate-50 dark:bg-slate-900">
+      </head>      <body className="relative min-h-screen overflow-y-auto bg-slate-50 dark:bg-slate-900 flex flex-col">
         <Navbar locale={params.lang} user={user} />
         <Content>{children}</Content>
+        <FooterWrapper locale={params.lang} />
       </body>
     </html>
   );